src/python/requirements.txt
mysql-connector-python
simplekml